+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v2.0.2
+
+ - Added `copyAsSnippetObj` and `copyAsSnippetBody` command.
+ - removed leading and trailing newline in copy text.
+
 ## v2.0.0, v2.0.1
 
 - Added Keybindings. Press `Ctrl+Shift+C` to Copy as Snippet.